Restructure linop.xray.astra
