+++ scico/test/linop/xray/astra/test_astra_2d.py
@@ -1,0 +1,187 @@
+import numpy as np
+
+import jax
+
+import pytest
+
+import scico
+import scico.numpy as snp
+from scico.linop import DiagonalStack
+from scico.test.linop.test_linop import adjoint_test
+
+try:
+    from scico.linop.xray.astra import (
+        XRayTransform2D,
+    )
+except ModuleNotFoundError as e:
+    if e.name == "astra":
+        pytest.skip("astra not installed", allow_module_level=True)
+    else:
+        raise e
+
+
+N = 128
+RTOL_CPU = 1e-4
+RTOL_GPU = 1e-1
+RTOL_GPU_RANDOM_INPUT = 2.0
+
+
+def make_im(Nx, Ny, is_3d=True):
+    x, y = snp.meshgrid(snp.linspace(-1, 1, Nx), snp.linspace(-1, 1, Ny), indexing="ij")
+
+    im = snp.where((x - 0.25) ** 2 / 3 + y**2 < 0.1, 1.0, 0.0)
+    if is_3d:
+        im = im[snp.newaxis, :, :]
+    im = im.astype(snp.float32)
+
+    return im
+
+
+def get_tol():
+    if jax.devices()[0].device_kind == "cpu":
+        rtol = RTOL_CPU
+    else:
+        rtol = RTOL_GPU  # astra inaccurate in GPU
+    return rtol
+
+
+def get_tol_random_input():
+    if jax.devices()[0].device_kind == "cpu":
+        rtol = RTOL_CPU
+    else:
+        rtol = RTOL_GPU_RANDOM_INPUT  # astra more inaccurate in GPU for random inputs
+    return rtol
+
+
+class XRayTransform2DTest:
+    def __init__(self, volume_geometry):
+        N_proj = 180  # number of projection angles
+        N_det = 384
+        det_spacing = 1
+        angles = np.linspace(0, np.pi, N_proj, False)
+
+        np.random.seed(1234)
+        self.x = np.random.randn(N, N).astype(np.float32)
+        self.y = np.random.randn(N_proj, N_det).astype(np.float32)
+        self.A = XRayTransform2D(
+            input_shape=(N, N),
+            det_count=N_det,
+            det_spacing=det_spacing,
+            angles=angles,
+            volume_geometry=volume_geometry,
+        )
+
+
+@pytest.fixture(params=[None, [-N / 2, N / 2, -N / 2, N / 2]])
+def testobj(request):
+    yield XRayTransform2DTest(request.param)
+
+
+def test_init(testobj):
+    with pytest.raises(ValueError):
+        A = XRayTransform2D(
+            input_shape=(16, 16, 16),
+            det_count=16,
+            det_spacing=1.0,
+            angles=np.linspace(0, np.pi, 32, False),
+        )
+    with pytest.raises(ValueError):
+        A = XRayTransform2D(
+            input_shape=(16, 16),
+            det_count=16.3,
+            det_spacing=1.0,
+            angles=np.linspace(0, np.pi, 32, False),
+        )
+    with pytest.raises(ValueError):
+        A = XRayTransform2D(
+            input_shape=(16, 16),
+            det_count=16,
+            det_spacing=1.0,
+            angles=np.linspace(0, np.pi, 32, False),
+            device="invalid",
+        )
+
+
+def test_2d_det_offset():
+    x = np.zeros((32, 32), dtype=np.float32)
+    x[8:-8, 8:-8] = 1.0
+    A = XRayTransform2D(x.shape, 40, 1.0, np.linspace(0, np.pi, 90))
+    shift = 4
+    As = XRayTransform2D(x.shape, 40, 1.0, np.linspace(0, np.pi, 90), shift)
+    y = A(x)
+    ys = As(x)
+    yss = np.roll(ys, shift, axis=1)
+    np.testing.assert_almost_equal(yss, y, decimal=4)
+
+
+def test_ATA_call(testobj):
+    # Test for the call-based interface
+    Ax = testobj.A(testobj.x)
+    ATAx = testobj.A.adj(Ax)
+    np.testing.assert_allclose(np.sum(testobj.x * ATAx), np.linalg.norm(Ax) ** 2, rtol=get_tol())
+
+
+def test_ATA_matmul(testobj):
+    # Test for the matmul interface
+    Ax = testobj.A @ testobj.x
+    ATAx = testobj.A.T @ Ax
+    np.testing.assert_allclose(np.sum(testobj.x * ATAx), np.linalg.norm(Ax) ** 2, rtol=get_tol())
+
+
+def test_AAT_call(testobj):
+    # Test for the call-based interface
+    ATy = testobj.A.adj(testobj.y)
+    AATy = testobj.A(ATy)
+    np.testing.assert_allclose(np.sum(testobj.y * AATy), np.linalg.norm(ATy) ** 2, rtol=get_tol())
+
+
+def test_AAT_matmul(testobj):
+    # Test for the matmul interface
+    ATy = testobj.A.T @ testobj.y
+    AATy = testobj.A @ ATy
+    np.testing.assert_allclose(np.sum(testobj.y * AATy), np.linalg.norm(ATy) ** 2, rtol=get_tol())
+
+
+def test_grad(testobj):
+    # ensure that we can take grad on a function using our projector
+    # grad || A(x) ||_2^2 == 2 A.T @ A x
+    A = testobj.A
+    x = testobj.x
+    g = lambda x: jax.numpy.linalg.norm(A(x)) ** 2
+    np.testing.assert_allclose(
+        scico.grad(g)(x), 2 * A.adj(A(x)), atol=get_tol() * x.max(), rtol=get_tol()
+    )
+
+
+def test_adjoint_grad(testobj):
+    A = testobj.A
+    x = testobj.x
+    Ax = A @ x
+    f = lambda y: jax.numpy.linalg.norm(A.T(y)) ** 2
+    np.testing.assert_allclose(scico.grad(f)(Ax), 2 * A(A.adj(Ax)), rtol=get_tol())
+
+
+def test_adjoint_random(testobj):
+    A = testobj.A
+    adjoint_test(A, rtol=10 * get_tol_random_input())
+
+
+def test_adjoint_typical_input(testobj):
+    A = testobj.A
+    x = make_im(A.input_shape[0], A.input_shape[1], is_3d=False)
+
+    adjoint_test(A, x=x, rtol=get_tol())
+
+
+def test_fbp(testobj):
+    x = testobj.A.fbp(testobj.y)
+    # Test for a bug (related to calling the Astra CPU FBP implementation
+    # when using a GPU device) that resulted in a constant zero output.
+    assert np.sum(np.abs(x)) > 0.0
+
+
+def test_jit_in_DiagonalStack():
+    """See https://github.com/lanl/scico/issues/331"""
+    N = 10
+    H = DiagonalStack([XRayTransform2D((N, N), N, 1.0, snp.linspace(0, snp.pi, N))])
+    H.T @ snp.zeros(H.output_shape, dtype=snp.float32)
